Configure next js for github pages
We want to host this portfolio on github pages.
To be able to do this we need to change the
configuration to support static exports.
We need to set the output to export, set
the base path and disable automatic image
optimization since it is not supported in github
pages.